算了，搞不来

+++ 强化学习/common/total_seconds.py
@@ -23,7 +23,7 @@ def total_seconds(time_start):
     # 计算分钟数
     minutes = int((total_seconds % 3600) // 60)
     # 计算秒数
-    seconds = int(total_seconds % 60)
+    seconds = float(total_seconds % 60)
     # 输出到日志
     logger.debug(f"用时: {hours}小时 {minutes}分钟 {seconds}秒")
